[ATIP] Move common funcs/steps to common folder and rename improper func name
