+++ tools/atip/atip/android/steps.py
@@ -33,10 +33,6 @@ reload(sys)
 sys.setdefaultencoding('utf-8')
 
 
-# @step(u'launch "{app_name}" on android')
-# def launch_app_by_name(context, app_name):
-#     android.launch_app_by_name(context, app_name)
-
 @step(u'I launch "{app_name}" with "{apk_pkg_name}" and "{apk_activity_name}" on android')
 def launch_app_by_names(context, app_name, apk_pkg_name, apk_activity_name):
     android.launch_app_by_name(
@@ -44,49 +40,6 @@ def launch_app_by_names(context, app_name, apk_pkg_name, apk_activity_name):
         app_name,
         apk_pkg_name,
         apk_activity_name)
-
-
-@step(u'I turn on device')
-def wake_up(context):
-	context.android.turnOnDevice()
-
-
-@step(u'I turn off device')
-def wake_up(context):
-	context.android.turnOffDevice()	
-
-# The possible orientation is:
-# natural or n
-# left or l
-# right or r
-# upsidedown or u (notes: "upsidedown" can not be set until Android 4.3.)
-@step(u'I set orientation "{orientation}"')
-def set_orientation(context, orientation):
-	context.android.setDeviceOrientation(orientation)
-
-
-# take screenshot and save to local file "home.png", can not work until Android 4.2.
-@step(u'I take screenshot as "{name}"')
-def take_screenshot(context, name):
-	context.android.takeScreenshot(name)
-
-
-# open notification, can not work until Android 4.3.
-@step(u'I open notification')
-def open_notification(context):
-	assert context.android.openNotification()
-
-
-# open quick settings, can not work until Android 4.3.
-@step(u'I open quick settings')
-def open_quick_settings(context):
-	assert context.android.openQuickSettings()
-
-# frequently-used key: home, back, left, right, up, down, center, menu, search, enter, 
-# delete(or del), recent(recent apps), volume_up, volume_down, volume_mute, camera, power
-@step(u'I press "{key}" on android')
-def press_key(context, key):
-	context.android.pressKeyBy(key)
 
 
 @step(u'I force to run all watchers')
@@ -154,25 +107,25 @@ def select_relative_any_object(context, class_target, relative, class_name, view
 def click_button_object(context, button_name):
 	ob = context.android.selectBtnObjectBy(button_name)
 	if ob.exists:
-		assert context.android.clickBtnObject(ob)
+		assert context.android.clickObject(ob)
 	else:
 		ob = context.android.selectImageBtnObjectBy(button_name)
 		assert ob.exists
-		assert context.android.clickBtnObject(ob)
+		assert context.android.clickObject(ob)
 
 
 @step(u'I click other "{class_name}" by "{which_key}" "{which_value}"')
 def click_other_view(context, class_name, which_key, which_value):
 	ob = context.android.selcetObjectBy(which_key, which_value, class_name)
 	assert ob.exists
-	assert context.android.clickBtnObject(ob)
+	assert context.android.clickObject(ob)
 
 
 @step(u'I click object "{key}"')
 def click_any_object(context, key):
 	ob = context.android.get2InfoTemp(key)
 	assert ob.exists
-	assert context.android.clickBtnObject(ob)
+	assert context.android.clickObject(ob)
 
 
 @step(u'I edit text "{edit_text}" to input "{text}"')
